+++ package.json
@@ -30,5 +30,8 @@
     "lerna": "^5.4.3",
     "prettier": "^2.8.8"
   },
-  "packageManager": "yarn@3.6.0"
+  "packageManager": "yarn@3.6.0",
+  "dependencies": {
+    "classnames": "^2.3.2"
+  }
 }
