+++ packages/client/src/pages/page-lobby/components/active-forum-topic/active-forum-topic.module.scss
@@ -4,15 +4,20 @@
   height: 100vh;
   left: 0;
   top: 0;
+
+  z-index: 1000;
+
   &.purple {
     background: rgba(44, 29, 66, 0.5);
   }
+
   &.neon {
     background: rgba(29, 66, 66, 0.5);
   }
-  z-index: 1000;
+
   .modal {
     position: relative;
+
     &_close {
       position: absolute;
       right: 35px;
@@ -22,6 +27,7 @@
 
       cursor: pointer;
     }
+
     &_wrapper {
       position: relative;
       display: block;
@@ -29,12 +35,15 @@
       height: 714px;
       margin: 5vh auto;
       border: 1px solid #FFFFFF;
+
       &_purple {
         background: #2C1D42;
       }
+
       &_neon {
         background: #1d4242;
       }
+
       &::before {
         content: url("../../../../assets/images/window-bl-large.svg");
         position: absolute;
@@ -42,6 +51,7 @@
         bottom: -15px;
         pointer-events: none;
       }
+
       &::after {
         content: url("../../../../assets/images/window-rt-large.svg");
         position: absolute;
@@ -50,6 +60,7 @@
         pointer-events: none;
       }
     }
+
     &_header {
       width: 467px;
       text-align: center;
@@ -61,6 +72,7 @@
       text-transform: uppercase;
     }
   }
+
   .topic {
     display: block;
     width: 738px;
@@ -69,12 +81,15 @@
     padding: 18px 24px;
     margin: 0 auto;
     margin: 15px 50px;
+
     &_purple {
       background: #BB6AEA30;
     }
+
     &_neon {
       background: #71AEA030;
     }
+
     &_descr {
       position: relative;
       color: #FFF;
@@ -82,6 +97,7 @@
       font-family: Blender Pro;
       font-weight: 500;
       line-height: 100%;
+
       &::after {
         position: absolute;
         bottom: -10px;
@@ -92,6 +108,7 @@
         background: #FFFFFF50;
       }
     }
+
     &_empty {
       color: rgba(255, 255, 255, 0.80);
       font-size: 15px;
@@ -101,6 +118,7 @@
     }
   }
 }
+
 .flex_column {
   display: flex;
   flex-direction: column;
