+++ packages/client/src/pages/page-lobby/components/active-forum-topic/components/topic-subcomment/topic-subcomment.module.scss
@@ -1,4 +1,5 @@
 .subcomment {
+
   &_wrapper {
     margin-top: 10px;
 
@@ -7,6 +8,7 @@
     justify-content: start;
     gap: 10px;
   }
+
   &_avatar {
     width: 33px;
     height: 33px;
@@ -14,7 +16,9 @@
     border: 1px solid #FFF;
     box-shadow: 0px 0px 10px 0px rgba(255, 255, 255, 0.50);
   }
+
   &_header {
+
     .name {
       text-align: right;
       font-size: 14px;
@@ -22,13 +26,16 @@
       text-transform: uppercase;
 
       margin-right: 5px;
+
       &_neon {
         color: #7CF0F0;
       }
+
       &_purple {
         color: #E2B0FF;
       }
     }
+
     .time {
       color: rgba(255, 255, 255, 0.80);
       font-size: 15px;
