+++ packages/client/src/pages/page-lobby/components/forum-item/forum-item.module.scss
@@ -1,5 +1,6 @@
 .forum {
   position: relative;
+
   &_wrapper {
     width: 100%;
     height: 115px;
@@ -13,29 +14,35 @@
 
     &.purpur {
       background: rgba(187, 106, 234, 0.20);
+
       &:hover {
         background: rgba(187, 106, 234, 0.40);
       }
     }
+
     &.neon {
       background: rgba(63, 183, 234, 0.20);
+
       &:hover {
         background: rgba(63, 183, 234, 0.40);
       ;
       }
     }
   }
+
   &_header {
     display: flex;
     flex-direction: row;
     justify-content: space-between;
     align-items: center;
+
     .title {
       color: #FFF;
       font-size: 14px;
       font-family: New_Zelek;
       text-transform: uppercase;
     }
+
     .date {
       color: rgba(255, 255, 255, 0.80);
       font-size: 15px;
@@ -43,6 +50,7 @@
       font-weight: 500;
     }
   }
+
   &_body {
     display: flex;
     flex-direction: row;
@@ -50,6 +58,7 @@
     align-items: flex-start;
 
     margin-top: 15px;
+
     .image {
       width: 40px;
       height: 40px;
@@ -59,11 +68,13 @@
 
       margin-right: 12px;
     }
+
     .message {
       font-size: 15px;
       font-weight: 500;
     }
   }
+
   &_open {
     display: inline;
     border: none;
