+++ packages/client/src/pages/page-lobby/components/rating/rating.tsx
@@ -3,11 +3,8 @@
 import styles from './rating.module.scss'
 import { useTheme } from '@hooks/use-theme'
 import { Theme } from '@config/constants'
-
-const mock_rating = {
-  score: 111,
-  place: 11,
-}
+import { mock_rating } from '@pages/page-lobby/mocks'
+import classNames from 'classnames'
 
 export const Rating = () => {
   // const { rating } = useSelector((store: RootState) => store)
@@ -16,9 +13,10 @@ export const Rating = () => {
   return (
     <div className={styles.rating}>
       <h3
-        className={`${styles.rating_title} ${
+        className={classNames(
+          styles.rating_title,
           themeName === Theme.Purple ? styles.purpur : styles.neon
-        }`}>
+        )}>
         рейтинг
       </h3>
       <span className={styles.rating_place}>
