+++ packages/client/src/pages/page-lobby/components/table-item/table-item.tsx
@@ -1,10 +1,11 @@
-import { useEffect, useState } from 'react'
+import { useMemo } from 'react'
 import { User } from '../../../../core/config/user-types'
 import userM from '../../../../assets/images/user-m.png'
 import styles from './table-item.module.scss'
 import { NavLink } from 'react-router-dom'
 import { PATH, Theme } from '../../../../core/config/constants'
 import { useTheme } from '@hooks/use-theme'
+import classNames from 'classnames'
 
 interface ITableItemProps {
   id: number
@@ -18,7 +19,6 @@ export const TableItem: React.FC<ITableItemProps> = ({
   users,
   password,
 }) => {
-  const [usersImg, setUsersImg] = useState<string[]>(getUsersImg(users))
   const { themeName } = useTheme()
 
   function getUsersImg(users: User[]) {
@@ -29,15 +29,14 @@ export const TableItem: React.FC<ITableItemProps> = ({
     })
   }
 
-  useEffect(() => {
-    setUsersImg(getUsersImg(users))
-  }, [users])
+  const usersImg = useMemo(() => getUsersImg(users), [users])
 
   return (
     <div
-      className={`${styles.table} ${
+      className={classNames(
+        styles.table,
         themeName === Theme.Purple ? styles.purpur : styles.neon
-      }`}>
+      )}>
       <h3 className={styles.table_title}>{title}</h3>
       <div className={styles.table_users}>
         {usersImg.map(user => {
@@ -50,17 +49,18 @@ export const TableItem: React.FC<ITableItemProps> = ({
           )
         })}
       </div>
-      {password ? (
+      {password && (
         <input
           type={'password'}
           placeholder={'введите пароль'}
           className={styles.table_password}
         />
-      ) : null}
+      )}
       <NavLink
-        className={`nav_link mini ${styles.table_sign} ${
+        className={classNames(
+          styles.table_sign,
           themeName === Theme.Purple ? styles.purple : styles.neon
-        } `}
+        )}
         to={`${PATH.GAME}`}>
         зайти
       </NavLink>
