+++ packages/client/src/pages/page-lobby/page-lobby.tsx
@@ -2,7 +2,7 @@ import styles from './page-lobby.module.scss'
 import { useSelector } from 'react-redux'
 import { TableItem } from './components/table-item/table-item'
 import React, { BaseSyntheticEvent, useEffect, useState } from 'react'
-import { ForumItem, IChatData } from './components/forum-item/forum-item'
+import { ForumItem } from './components/forum-item/forum-item'
 import avatar from '../../assets/images/chat-avatar.png'
 import { RootState } from '@store/store'
 import { useTheme } from '@hooks/use-theme'
@@ -11,46 +11,9 @@ import settings from '@images/settings.svg'
 import { Rating } from '@pages/page-lobby/components/rating/rating'
 import { BattleSetting } from '@pages/page-lobby/components/battle-settings/battle-settings'
 import { ActiveTopicModal } from '@pages/page-lobby/components/active-forum-topic/active-forum-topic'
-
-const tables_mock = [
-  {
-    id: 832,
-    title: 'string',
-    users: [
-      {
-        id: 111,
-        first_name: 'string',
-        second_name: 'string',
-        display_name: 'string',
-        login: 'string',
-        email: 'string',
-        phone: 'string',
-        avatar: '',
-      },
-    ],
-    // password: 111
-  },
-]
-const forum_mock_topics: IChatData[] | [] = [
-  {
-    id: 123,
-    title: 'my-chat',
-    avatar: null,
-    unread_count: 15,
-    last_message: {
-      user: {
-        first_name: 'Petya',
-        second_name: 'Pupkin',
-        avatar: '',
-        email: 'my@email.com',
-        login: 'userLogin',
-        phone: '8(911)-222-33-22',
-      },
-      time: '2020-01-02T14:22:22.000Z',
-      content: 'this is message content',
-    },
-  },
-]
+import { forum_mock_topics, tables_mock } from '@pages/page-lobby/mocks'
+import classNames from 'classnames'
+import { IChatData } from '@pages/page-lobby/types'
 
 export const PageLobby = () => {
   const { user } = useSelector((store: RootState) => store)
@@ -61,10 +24,10 @@ export const PageLobby = () => {
   const [activeTopicId, setActiveTopicId] = useState(0)
 
   useEffect(() => {
-    OnSearchInput(searchString, topicList)
+    onSearchInput(searchString, topicList)
   }, [searchString])
 
-  function OnSearchInput(value: string, array: IChatData[]) {
+  function onSearchInput(value: string, array: IChatData[]) {
     if (Array.isArray(array) && array[0] && value) {
       const res = array.filter(chat => chat.title.includes(value))
       setTopicList(res)
@@ -76,9 +39,10 @@ export const PageLobby = () => {
   return (
     <div className={styles.container}>
       <section
-        className={`${styles.window} ${
+        className={classNames(
+          styles.window,
           themeName === Theme.Purple ? styles.purpur : styles.neon
-        }`}>
+        )}>
         <h3 className={styles.window_header}>доступные битвы</h3>
         <div className={styles.window_wrapper}>
           {tables_mock.map(table => {
@@ -87,27 +51,30 @@ export const PageLobby = () => {
         </div>
       </section>
       <section
-        className={`${styles.forum} ${
+        className={classNames(
+          styles.forum,
           themeName === Theme.Purple ? styles.purpur : styles.neon
-        }`}>
+        )}>
         <input
           type={'search'}
           onChange={(event: BaseSyntheticEvent) =>
             setSearchString(event.target.value)
           }
           placeholder={'Поиск по названию'}
-          className={`${styles.forum_search} ${
+          className={classNames(
+            styles.forum_search,
             themeName === Theme.Purple ? styles.purpur : styles.neon
-          }`}
+          )}
         />
         {topicList.map(topic => {
           return <ForumItem {...topic} onClick={setActiveTopicId} />
         })}
         <div className={styles.forum_input_wrapper}>
           <input
-            className={`${styles.forum_input} ${
+            className={classNames(
+              styles.forum_input,
               themeName === Theme.Purple ? styles.purpur : styles.neon
-            }`}
+            )}
             onBlur={(event: BaseSyntheticEvent) => {
               console.log('new topic name', event.target.value)
               setNewTopic(event.target.value)
@@ -120,19 +87,21 @@ export const PageLobby = () => {
             onClick={() => {
               console.log('topic added')
             }}
-            className={`${styles.forum_input_submit} ${
+            className={classNames(
+              styles.forum_input_submit,
               themeName === Theme.Purple
                 ? styles.forum_input_submit_purpur
                 : styles.forum_input_submit_neon
-            }`}>
+            )}>
             Создать тему
           </button>
         </div>
       </section>
       <section
-        className={`${styles.user} ${
+        className={classNames(
+          styles.user,
           themeName === Theme.Purple ? styles.user_purpur : styles.user_neon
-        }`}>
+        )}>
         <img
           src={
             user.user?.avatar
@@ -151,15 +120,17 @@ export const PageLobby = () => {
           </a>
         </div>
         <a
-          className={`${styles.user_link} ${
+          className={classNames(
+            styles.user_link,
             themeName === Theme.Purple ? styles.purpur : styles.neon
-          }`}>
+          )}>
           как играть?
         </a>
         <a
-          className={`${styles.user_link} ${
+          className={classNames(
+            styles.user_link,
             themeName === Theme.Purple ? styles.purpur : styles.neon
-          }`}>
+          )}>
           бестиарий
         </a>
         <Rating />
