+++ packages/client/src/shared/ui/checkbox/checkbox.module.scss
@@ -2,13 +2,16 @@
   font-size: 14px;
   font-family: New_Zelek;
   margin-right: 10px;
+
   &.purpur {
     color: #DFADFF;
   }
+
   &.neon {
     color: #7CF0F0;
   }
 }
+
 .input {
   width: 10px;
   height: 10px;
