+++ packages/client/src/shared/ui/switcher/switcher.module.scss
@@ -1,6 +1,7 @@
 .wrapper {
   position: relative;
   margin: 0 auto;
+
   .switch {
     position: absolute;
     top: 1px;
@@ -10,6 +11,7 @@
     height: 11px;
     bottom: 14px;
   }
+
   &_label {
     width: 250px;
     padding-left: 0px;
@@ -20,26 +22,34 @@
     font-size: 14px;
     font-family: New_Zelek;
     line-height: 16px;
+
     .unactive {
+
       &_purpur {
         color: #E2B0FF65;
       }
+
       &_neon {
         color: #7CF0F065;
       }
     }
+
     .active {
       text-shadow: 0px 0px 5px 0px rgba(255, 255, 255, 0.50);
+
       &_purpur {
         color: #E2B0FF;
       }
+
       &_neon {
         color: #7CF0F0;
       }
     }
   }
+
   input {
     display: none;
+
     &:checked + .slider:before {
       -webkit-transform: translateX(9px);
       -ms-transform: translateX(9px);
@@ -47,6 +57,7 @@
     }
   }
 }
+
 .slider {
   position: absolute;
   cursor: pointer;
@@ -58,6 +69,7 @@
   -webkit-transition: .4s;
   transition: .4s;
   border: none;
+
   &:before {
     position: absolute;
     content: "";
@@ -70,14 +82,18 @@
     -webkit-transition: .4s;
     transition: .4s;
   }
+
   &_purpur {
     background-color: #BB6AEA66;
+
     &:before {
       background-color: #BB6AEA;
     }
   }
+
   &_neon {
     background-color: #7CF0F066;
+
     &:before {
       background-color: #7CF0F0;
     }
